refactor(entail): split reasoner into vocab/interner/rdfs modules + Regime::Rif scaffolding
Split the flat lib.rs into a module tree (vocab, interner, rdfs) with
lib.rs as a thin facade. No behavior change: the RDFS chase logic moved
verbatim into rdfs.rs.

Add API scaffolding for the forthcoming OWL-Direct and RIF engines:
- Regime::Rif variant plus Regime::from_iri("RIF")
- EntailError::Parse and EntailError::Inconsistent variants

All 30 conformance Entailment xfails unchanged; make check and make wasm
green.

Refs #37
